Implement creating a Schema object from JsonElement
