+++ src/Mscc.GenerativeAI/Types/Generative/Schema.cs
@@ -10,7 +10,7 @@ namespace Mscc.GenerativeAI
     /// <summary>
     /// The Schema object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. Represents a select subset of an OpenAPI 3.0 schema object.
     /// </summary>
-    public class Schema
+    public partial class Schema
     {
         /// <summary>
         /// Required. Data type.
@@ -68,7 +68,7 @@ namespace Mscc.GenerativeAI
         /// Optional. Properties of Type.OBJECT.
         /// An object containing a list of "key": value pairs. Example: { "name": "wrench", "mass": "1.3kg", "count": "3" }.
         /// </summary>
-        public dynamic? Properties { get; set; }
+        public object? Properties { get; set; }
 
         /// <summary>
         /// Optional. The order of the properties. Not a standard field in open api spec. Used to determine the order of the properties in the response.
@@ -132,303 +132,5 @@ namespace Mscc.GenerativeAI
         /// Optional. Pattern of the Type.STRING to restrict a string to a regular expression.
         /// </summary>
         public string? Pattern { get; set; }
-
-        /// <summary>
-        /// Converts a JsonElement representing a JSON Schema to a Schema object.
-        /// </summary>
-        /// <param name="jsonElement">The JsonElement containing the JSON Schema</param>
-        /// <returns>A Schema object representing the JSON Schema</returns>
-        public static Schema FromJsonElement(JsonElement jsonElement)
-        {
-            var schema = new Schema();
-
-            if (jsonElement.ValueKind != JsonValueKind.Object)
-            {
-                throw new ArgumentException("JSON Schema must be an object", nameof(jsonElement));
-            }
-
-            // Parse type (can be string or array)
-            if (jsonElement.TryGetProperty("type", out var typeElement))
-            {
-                var (parameterType, isNullable) = ParseParameterType(typeElement);
-                schema.Type = parameterType;
-
-                // Set nullable if "null" was found in type array, but don't override explicit nullable property
-                if (isNullable && !jsonElement.TryGetProperty("nullable", out _))
-                {
-                    schema.Nullable = true;
-                }
-            }
-
-            // Parse basic properties
-            if (jsonElement.TryGetProperty("title", out var titleElement))
-            {
-                schema.Title = titleElement.GetString();
-            }
-
-            if (jsonElement.TryGetProperty("format", out var formatElement))
-            {
-                schema.Format = formatElement.GetString() ?? "";
-            }
-
-            if (jsonElement.TryGetProperty("description", out var descriptionElement))
-            {
-                schema.Description = descriptionElement.GetString() ?? "";
-            }
-
-            // Explicit nullable property takes precedence over type array inference
-            if (jsonElement.TryGetProperty("nullable", out var nullableElement))
-            {
-                schema.Nullable = nullableElement.GetBoolean();
-            }
-
-            if (jsonElement.TryGetProperty("default", out var defaultElement))
-            {
-                schema.Default = ParseValue(defaultElement);
-            }
-
-            if (jsonElement.TryGetProperty("example", out var exampleElement))
-            {
-                schema.Example = ParseValue(exampleElement);
-            }
-
-            // Parse array-specific properties
-            if (jsonElement.TryGetProperty("items", out var itemsElement))
-            {
-                schema.Items = FromJsonElement(itemsElement);
-            }
-
-            if (jsonElement.TryGetProperty("maxItems", out var maxItemsElement))
-            {
-                schema.MaxItems = maxItemsElement.GetInt64();
-            }
-
-            if (jsonElement.TryGetProperty("minItems", out var minItemsElement))
-            {
-                schema.MinItems = minItemsElement.GetInt64();
-            }
-
-            // Parse string-specific properties
-            if (jsonElement.TryGetProperty("enum", out var enumElement))
-            {
-                schema.Enum = ParseStringArray(enumElement);
-            }
-
-            if (jsonElement.TryGetProperty("maxLength", out var maxLengthElement))
-            {
-                schema.MaxLength = maxLengthElement.GetString();
-            }
-
-            if (jsonElement.TryGetProperty("minLength", out var minLengthElement))
-            {
-                schema.MinLength = minLengthElement.GetString();
-            }
-
-            if (jsonElement.TryGetProperty("pattern", out var patternElement))
-            {
-                schema.Pattern = patternElement.GetString();
-            }
-
-            // Parse number-specific properties
-            if (jsonElement.TryGetProperty("maximum", out var maximumElement))
-            {
-                schema.Maximum = maximumElement.GetInt64();
-            }
-
-            if (jsonElement.TryGetProperty("minimum", out var minimumElement))
-            {
-                schema.Minimum = minimumElement.GetInt64();
-            }
-
-            // Parse object-specific properties
-            if (jsonElement.TryGetProperty("properties", out var propertiesElement))
-            {
-                schema.Properties = ParseProperties(propertiesElement);
-            }
-
-            if (jsonElement.TryGetProperty("required", out var requiredElement))
-            {
-                schema.Required = ParseStringArray(requiredElement);
-            }
-
-            if (jsonElement.TryGetProperty("propertyOrdering", out var propertyOrderingElement))
-            {
-                schema.PropertyOrdering = ParseStringArray(propertyOrderingElement);
-            }
-
-            if (jsonElement.TryGetProperty("maxProperties", out var maxPropertiesElement))
-            {
-                schema.MaxProperties = maxPropertiesElement.GetString();
-            }
-
-            if (jsonElement.TryGetProperty("minProperties", out var minPropertiesElement))
-            {
-                schema.MinProperties = minPropertiesElement.GetString();
-            }
-
-            // Parse anyOf
-            if (jsonElement.TryGetProperty("anyOf", out var anyOfElement))
-            {
-                schema.AnyOf = ParseSchemaArray(anyOfElement);
-            }
-
-            return schema;
-        }
-
-        private static (ParameterType, bool) ParseParameterType(JsonElement typeElement)
-        {
-            if (typeElement.ValueKind == JsonValueKind.String)
-            {
-                // Single type as string
-                var type = ParseSingleParameterType(typeElement.GetString());
-                return (type, false);
-            }
-            else if (typeElement.ValueKind == JsonValueKind.Array)
-            {
-                // Array of types
-                var types = new List<string>();
-                foreach (var item in typeElement.EnumerateArray())
-                {
-                    if (item.ValueKind == JsonValueKind.String)
-                    {
-                        var typeStr = item.GetString();
-                        if (typeStr != null)
-                        {
-                            types.Add(typeStr);
-                        }
-                    }
-                }
-
-                if (types.Count == 0)
-                {
-                    return (ParameterType.TypeUnspecified, false);
-                }
-
-                // Check if "null" is present
-                bool isNullable = types.Contains("null");
-
-                // Get the first non-null type as the actual type
-                var actualType = types.FirstOrDefault(t => t != "null");
-                var parameterType = ParseSingleParameterType(actualType);
-
-                return (parameterType, isNullable);
-            }
-            else
-            {
-                return (ParameterType.TypeUnspecified, false);
-            }
-        }
-
-        private static ParameterType ParseSingleParameterType(string? typeString)
-        {
-            return typeString?.ToLowerInvariant() switch
-            {
-                "string" => ParameterType.String,
-                "number" => ParameterType.Number,
-                "integer" => ParameterType.Integer,
-                "boolean" => ParameterType.Boolean,
-                "array" => ParameterType.Array,
-                "object" => ParameterType.Object,
-                "null" => ParameterType.Null,
-                _ => ParameterType.TypeUnspecified
-            };
-        }
-
-        private static object? ParseValue(JsonElement element)
-        {
-            return element.ValueKind switch
-            {
-                JsonValueKind.String => element.GetString(),
-                JsonValueKind.Number => element.TryGetInt64(out var longValue) ? longValue : element.GetDouble(),
-                JsonValueKind.True => true,
-                JsonValueKind.False => false,
-                JsonValueKind.Null => null,
-                JsonValueKind.Array => ParseArray(element),
-                JsonValueKind.Object => ParseObject(element),
-                _ => null
-            };
-        }
-
-        private static List<object> ParseArray(JsonElement arrayElement)
-        {
-            var result = new List<object>();
-            foreach (var item in arrayElement.EnumerateArray())
-            {
-                var value = ParseValue(item);
-                if (value != null)
-                {
-                    result.Add(value);
-                }
-            }
-
-            return result;
-        }
-
-        private static Dictionary<string, object> ParseObject(JsonElement objectElement)
-        {
-            var result = new Dictionary<string, object>();
-            foreach (var property in objectElement.EnumerateObject())
-            {
-                var value = ParseValue(property.Value);
-                if (value != null)
-                {
-                    result[property.Name] = value;
-                }
-            }
-
-            return result;
-        }
-
-        private static List<string>? ParseStringArray(JsonElement arrayElement)
-        {
-            if (arrayElement.ValueKind != JsonValueKind.Array)
-            {
-                return null;
-            }
-
-            var result = new List<string>();
-            foreach (var item in arrayElement.EnumerateArray())
-            {
-                var value = item.GetString();
-                if (value != null)
-                {
-                    result.Add(value);
-                }
-            }
-
-            return result.Count > 0 ? result : null;
-        }
-
-        private static List<Schema>? ParseSchemaArray(JsonElement arrayElement)
-        {
-            if (arrayElement.ValueKind != JsonValueKind.Array)
-            {
-                return null;
-            }
-
-            var result = new List<Schema>();
-            foreach (var item in arrayElement.EnumerateArray())
-            {
-                result.Add(FromJsonElement(item));
-            }
-
-            return result.Count > 0 ? result : null;
-        }
-
-        private static dynamic? ParseProperties(JsonElement propertiesElement)
-        {
-            if (propertiesElement.ValueKind != JsonValueKind.Object)
-            {
-                return null;
-            }
-
-            var properties = new Dictionary<string, Schema>();
-            foreach (var property in propertiesElement.EnumerateObject())
-            {
-                properties[property.Name] = FromJsonElement(property.Value);
-            }
-
-            return properties;
-        }
     }
 }
